HA proxy: forward BGP and Karaf SSH connections
The management vip of the HA proxy needs to also forward BGP
connections to port 179 and Karaf ssh connections to port 8101.
The BGP connections need to be forwarded in order for OpenDaylight to
peer with an internal BGP router, possibly acting as the DC gateway.
The SSH forwarding is required so that the OpenDaylight BGP router can
be configured from the karaf shell. This is enabled with some
reservation, and will be disabled in the future once it is possible to
execute the required commands via rest calls.

Signed-off-by: Romanos Skiadas <rski@intracom-telecom.com>